Show the Pro settings page without the ten-click unlock
AI tasks were hidden twice over: behind an active license, and behind a
Pro settings page that only appeared after ten clicks on the version row.
The second gate was for a feature still in preparation, and it now only
keeps buyers from finding the form that takes their code.

Drop ProUnlockState and the click counter with it. The page is drawn
whenever the platform can run what the license unlocks, i.e. on desktop;
mobile still omits it from the definitions entirely so its contents stay
out of the settings search. The license gate itself is untouched.

README, both languages: document that the plugin talks to the license API
to verify an activation code, and what that request carries. The AI task
section claimed the plugin "makes no network calls itself", which has not
been true since license activation landed.
