+++ README.ja.md
@@ -73,6 +73,7 @@ scheduled_time: "09:00"
 - 実行ログスナップショットのバックアップ間隔/保持期間
 - カスタム時間帯境界と時間帯折りたたみUI
 - Google Calendar エクスポート既定値
+- Pro設定：ライセンスの有効化、デバイスのシート管理、それによって使えるAIタスクの各種設定（デスクトップのみ）
 
 現行コードのデフォルト値:
 
@@ -96,8 +97,17 @@ scheduled_time: "09:00"
 タスクをAI CLI（Claude Code / Codex）で実行し、タスクリスト下の「AI実行」ペインに埋め込まれた本物のターミナルで操作できます。対話型TUIがそのまま描画され、タスクノートのプロンプトが自動投入されたあとも、自分のターミナルでCLIを起動したときと同じように、いつでもセッションに直接入力できます。
 
 - `TaskChute settings` → `AIタスク` → `AIタスクを有効化` でオンにします。デフォルトはオフで、モバイルでは動作しません。
-- プラグインは手元にあるCLIを起動するだけです。**Claude Code（`claude`）や Codex（`codex`）のインストールと認証は、各自で事前に済ませておく必要があります。** プラグインはAPIキーを扱わず、自身でネットワーク通信も行いません。
+- AIタスクの利用には**Proライセンス**の有効化が必要です。`TaskChute settings` → `Pro設定` でアクティベーションコードを入力してください。同じページでシートを使用中のデバイスの一覧と解放も行えます。送信内容は[ネットワーク通信について](#ネットワーク通信について)を参照してください。
+- プラグインは手元にあるCLIを起動するだけです。**Claude Code（`claude`）や Codex（`codex`）のインストールと認証は、各自で事前に済ませておく必要があります。** プラグイン自身はAPIキーを扱わず、Vaultの内容をAIサービスへ送信することもありません。CLIは各自が認証したアカウントで、それぞれの提供元と通信します。
 - 自動検出に失敗する場合は、同じ設定セクションでバイナリの絶対パスを指定してください。
+
+### ネットワーク通信について
+
+プラグインがネットワーク通信を行うのは、Proライセンスの検証のためだけです。通信先はライセンスAPI `https://taskchute-license.levers.workers.dev`、経路はHTTPSです。
+
+- **通信のタイミング** — アクティベーションコードの登録時とデバイスのシート解放時、`Pro設定` ページでシート使用状況を一覧するとき（最短1分に1回まで）、そしてこのデバイスに保存された署名付きトークンの有効期限が近いとき（12時間ごとに確認、トークンの有効期間は7日）。更新の合間はトークンをオフラインで検証するため、接続がない環境でもそのまま利用できます。
+- **送信する内容** — ライセンスコード、このプラグインが生成したランダムなデバイスID、デバイスの一覧でシートを見分けるためのラベル（マシンのホスト名、取得できない場合はプラットフォーム名と、Vault名）、プラットフォーム名、プラグインのバージョン。
+- **送信しない内容** — タスク・ノート・ログ・プロンプトの内容は一切送信しません。テレメトリや利用状況の収集も行いません。上記以外にネットワーク通信を行う箇所はありません。
 
 タスクノートに frontmatter を付け、`## Prompt` セクションを追加します。
