+++ README.md
@@ -71,6 +71,7 @@ Open `TaskChute settings` to configure:
 - Backup interval/retention for execution snapshots
 - Custom time-slot boundaries and collapsible slot UI
 - Google Calendar export defaults
+- Pro settings: license activation, device seats, and the AI task options they unlock (desktop only)
 
 Current default values in code:
 
@@ -94,8 +95,17 @@ With default `vaultRoot` mode, TaskChute-managed folders are:
 Run a task with an AI CLI (Claude Code or Codex) inside a real terminal embedded in the "AI runs" pane below the task list: the full interactive TUI renders there, the prompt from the task note is submitted automatically, and you can type into the session at any time — exactly as if you had launched the CLI in your own terminal.
 
 - Enable it in `TaskChute settings` → `AI task` → `Enable AI tasks`. The feature is off by default and never activates on mobile.
-- The plugin launches the CLIs you already have: **Claude Code (`claude`) and/or Codex (`codex`) must be installed and authenticated separately by you.** The plugin handles no API keys and makes no network calls itself.
+- AI tasks need an activated **Pro license**. Enter your code in `TaskChute settings` → `Pro settings`; the same page lists the devices holding a seat and lets you release one. See [Network use](#network-use) for what activation sends.
+- The plugin launches the CLIs you already have: **Claude Code (`claude`) and/or Codex (`codex`) must be installed and authenticated separately by you.** The plugin handles no API keys of its own and sends nothing from your vault to any AI service — the CLI talks to its provider under the account you authenticated it with.
 - If auto-detection fails, set the absolute binary paths in the same settings section.
+
+### Network use
+
+The plugin makes network requests for exactly one purpose: verifying a Pro license against the license API at `https://taskchute-license.levers.workers.dev`, over HTTPS.
+
+- **When** — when you activate a code or release a device seat, when the `Pro settings` page lists the seats in use (throttled to at most one check a minute), and when the signed token stored on this device is close to expiring (checked every 12 hours; tokens are valid for 7 days). Between refreshes the token is verified offline, so a vault without a connection keeps working.
+- **What is sent** — the license code, a random device id generated by this plugin, a device label (your machine's hostname, or its platform name, together with the vault name) so seats are distinguishable in the device list, the platform name, and the plugin version.
+- **What is never sent** — no task, note, log or prompt content, and no telemetry or analytics of any kind. Nothing else in the plugin contacts the network.
 
 Mark a task note with frontmatter and add a `## Prompt` section:
