+++ src/i18n/locales/ja.ts
@@ -15,7 +15,6 @@ export const ja = {
     heading: "タスクシュートのファイルパス",
     version: {
       name: "バージョン",
-      proUnlocked: "Pro 設定を表示しました。",
     },
     storage: {
       baseLocationName: "デフォルトの場所",
