+++ tests/settings/settings-version.test.ts
@@ -1,7 +1,5 @@
-import type { SettingDefinitionAction } from 'obsidian'
-import { Notice, Platform, mockApp } from 'obsidian'
+import { mockApp } from 'obsidian'
 import { TaskChuteSettingTab } from '../../src/settings/SettingsTab'
-import { ProUnlockState, isProSectionVisible } from '../../src/settings/proUnlockState'
 import { versionSection } from '../../src/settings/sections/version'
 import type { SectionContext } from '../../src/settings/types'
 import { initializeLocaleManager, setLocaleOverride, t } from '../../src/i18n'
@@ -40,7 +38,7 @@ describe('TaskChute settings version display', () => {
   test('shows the version from manifest.json', () => {
     const ctx = createContext('1.7.10')
 
-    const row = findByName(versionSection(new ProUnlockState()).items(ctx), 'Version')
+    const row = findByName(versionSection().items(ctx), 'Version')
 
     expect(row?.desc).toBe('1.7.10')
   })
@@ -49,62 +47,10 @@ describe('TaskChute settings version display', () => {
     setLocaleOverride('ja')
     const ctx = createContext('1.7.11')
 
-    const items = versionSection(new ProUnlockState()).items(ctx)
+    const items = versionSection().items(ctx)
 
     expect(findByName(items, 'バージョン')?.desc).toBe('1.7.11')
     expect(t('settings.version.name', '__missing__')).not.toBe('__missing__')
-  })
-
-  test('unlocks the Pro section only after ten clicks', () => {
-    const ctx = createContext('1.7.12')
-    const unlock = new ProUnlockState()
-    const row = findByName(
-      versionSection(unlock).items(ctx),
-      'Version',
-    ) as SettingDefinitionAction
-    const click = () => row.action({} as HTMLElement, 0)
-
-    for (let i = 0; i < 9; i += 1) click()
-    expect(isProSectionVisible(ctx, unlock)).toBe(false)
-    // Only a visibility predicate changes, so the tree is never rebuilt.
-    expect(ctx.refreshDomState).not.toHaveBeenCalled()
-    expect(ctx.update).not.toHaveBeenCalled()
-
-    click()
-    expect(isProSectionVisible(ctx, unlock)).toBe(true)
-    expect(ctx.refreshDomState).toHaveBeenCalledTimes(1)
-    expect(Notice).toHaveBeenCalledTimes(1)
-
-    // Further clicks are inert once the section is already visible.
-    click()
-    expect(ctx.refreshDomState).toHaveBeenCalledTimes(1)
-  })
-
-  /**
-   * The section it would reveal is not declared on mobile, so counting the taps
-   * would end in a notice announcing settings that are not there.
-   */
-  test('does not unlock the Pro section on mobile', () => {
-    Platform.isDesktop = false
-    Platform.isMobile = true
-    try {
-      const ctx = createContext('1.7.13')
-      const unlock = new ProUnlockState()
-      const row = findByName(
-        versionSection(unlock).items(ctx),
-        'Version',
-      ) as SettingDefinitionAction
-
-      for (let i = 0; i < 12; i += 1) row.action({} as HTMLElement, 0)
-
-      expect(unlock.isUnlocked).toBe(false)
-      expect(isProSectionVisible(ctx, unlock)).toBe(false)
-      expect(Notice).not.toHaveBeenCalled()
-      expect(ctx.refreshDomState).not.toHaveBeenCalled()
-    } finally {
-      Platform.isDesktop = true
-      Platform.isMobile = false
-    }
   })
 
   test('the tab exposes the version row through its definitions', () => {
